+++ src/agents/investment/sub-agents/aquire/tools.ts
